+++ lib/slack-notifier/payload_middleware/stack.rb
@@ -25,10 +25,24 @@ module Slack
         end
 
         def call payload={}
-          stack.inject payload do |pld, middleware|
-            middleware.call(pld)
+          result = stack.inject payload do |pld, middleware|
+            as_array(pld).flat_map do |p|
+              middleware.call(p)
+            end
           end
+
+          as_array(result)
         end
+
+        private
+
+          def as_array args
+            if args.respond_to?(:to_ary)
+              args.to_ary
+            else
+              [args]
+            end
+          end
       end
     end
   end