+++ changelog.md
@@ -1,3 +1,5 @@
+- feat: support any middleware splittin payload into an array to allow multiple payloads from a single process.
+
 # 2.2.2
 - fix wrapping of attachments passed as a hash
 - fix error in `LinkFormatter` if a text payload was nil [#81]